requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
psycopg2-binary
